packaging: only advertise a single platform for compatibility.
A binary addon by definition only works on 1 platform. By filling this in at
build-time, it will make the add-ons easier to distribute via repositories,
because the platform string will sort out which repositories/add-ons to grab.

This is part of a quick effort to get Frodo bin-addon distribution up and
running. Each platform will have its own repository, so there's no need to try
to bundle various platforms into a single zip.

This was done blindly for osx/ios. For this to go in, we'll definitely need
sign-offs from osx/ios/windows devs. Also, windows will need to do some manual
operations here to fill in the platform and move it to the correct addon.xml.