change to use TinyXML2
The old version is no more maintained and comes the time to change to new.
TinyXML2 taken here is dated 7. June 2021, where much more actual as the old
where from 23. March 2015.

Also does maintainer about old ref to the new on http://www.grinninglizard.com/tinyxml/.

The changes on addon are not so much and on test it has worked with new.